requirements.txt
fastapi
uvicorn
pandas
numpy
mysql-connector-python
prophet
